+++ lib/src/article/add_all_calibres_and_articles.dart
@@ -4,13 +4,13 @@ import 'package:sembast/sembast.dart';
 // Project imports:
 import 'package:models_weebi/weebi_models.dart' show ArticleCalibre;
 import 'package:services_weebi/db_wrappers.dart';
-import 'package:services_weebi/src/article_no_sembast/add_all_calibres.dart';
+import 'package:services_weebi/src/article_no_sembast/create_all_calibres.dart';
 
 // this adds all but does not erase anything when importing
-class AddAllArticleCalibresRpc extends AddAllArticleCalibresAbstract {
+class CreateAllArticleCalibresRpc extends CreateAllArticleCalibresAbstract {
   final DbArticles _database;
 
-  const AddAllArticleCalibresRpc(this._database);
+  const CreateAllArticleCalibresRpc(this._database);
 
   @override
   Future<int> request(List<ArticleCalibre> calibres) async {