+++ lib/src/article_no_sembast/create_all_calibres.dart
@@ -6,14 +6,14 @@ import 'package:models_weebi/base.dart' show EndpointBase;
 
 // this adds all but does not erase anything when importing
 
-abstract class AddAllArticleCalibresAbstract
+abstract class CreateAllArticleCalibresAbstract
     implements EndpointBase<void, List<ArticleCalibre>> {
-  const AddAllArticleCalibresAbstract();
+  const CreateAllArticleCalibresAbstract();
   @override
   Future<void> request(List<ArticleCalibre> lines) async {}
 }
 
-class AddAllLineArticlesFakeRpc extends AddAllArticleCalibresAbstract {
+class AddAllLineArticlesFakeRpc extends CreateAllArticleCalibresAbstract {
   const AddAllLineArticlesFakeRpc();
 
   @override