+++ lib/src/db_store_refs.dart
@@ -6,6 +6,7 @@ class DbStoresWeebi extends DbStoresBase {
       intMapStoreFactory.store('herders'); // * this is not a mistaker
   static final config = stringMapStoreFactory.store('config'); // not used
   static final shop = intMapStoreFactory.store('shop');
+  static final articlePhotos = intMapStoreFactory.store('photos');
   static final oldProducts =
       intMapStoreFactory.store('products'); // only used for oldDb migration
   // do not be tempted to use or rename herders to contacts