+++ lib/src/tickets.dart
@@ -1,4 +1,5 @@
 // Project imports:
+import 'package:services_weebi/src/db_wrappers.dart';
 import 'package:services_weebi/src/ticket/add_all_tickets.dart';
 import 'package:services_weebi/src/ticket/add_ticket.dart';
 import 'package:services_weebi/src/ticket/delete_ticket.dart';
@@ -15,6 +16,7 @@ import 'package:services_weebi/src/ticket_no_sembast/get_tickets.dart';
 import 'package:services_weebi/src/ticket_no_sembast/restore_ticket.dart';
 
 abstract class TicketsServiceAbstract {
+  static const count = 7;
   final AddTicketAbstractRpc addTicketRpc;
   final DisableTicketAbstractRpc disableTicketRpc;
   final GetAllTicketsAbstractRpc getAllTicketsRpc;
@@ -34,34 +36,8 @@ abstract class TicketsServiceAbstract {
   );
 }
 
-class TicketsServiceNoSembast implements TicketsServiceAbstract {
-  const TicketsServiceNoSembast();
-
-  @override
-  AddAllTicketsAbstractRpc get addAllTicketsRpc => AddAllTicketsFakeRpc();
-
-  @override
-  AddTicketAbstractRpc get addTicketRpc => AddTicketFakeRpc();
-
-  @override
-  DeleteAllTicketsAbstractRpc get deleteAllTicketsRpc =>
-      DeleteAllTicketsFakeRpc();
-
-  @override
-  DeleteTicketAbstractRpc get deleteTicketRpc => DeleteTicketFakeRpc();
-
-  @override
-  DisableTicketAbstractRpc get disableTicketRpc => DisableTicketFakeRpc();
-
-  @override
-  GetAllTicketsAbstractRpc get getAllTicketsRpc => GetAllTicketsFakeRpc();
-
-  @override
-  RestoreTicketAbstractRpc get restoreTicketRpc => RestoreTicketFakeRpc();
-}
-
 class TicketsService extends TicketsServiceAbstract {
-  static final count = 7;
+  static const count = TicketsServiceAbstract.count;
 
   TicketsService(
     AddTicketRpc addTicketRpc,
@@ -80,4 +56,53 @@ class TicketsService extends TicketsServiceAbstract {
           deleteAllTicketsRpc,
           deleteTicketRpc,
         );
+}
+
+abstract class TicketsServiceInstantiater {
+  static get noPersistence => TicketsServiceNoSembast();
+
+  static TicketsService instTicketsStoreSembast(DbTickets db) {
+    final AddTicketRpc createTicketRpc = AddTicketRpc(db);
+    final DisableTicketRpc disableTicketRpc = DisableTicketRpc(db);
+    final GetAllTicketsRpc getTicketsRpc = GetAllTicketsRpc(db);
+    final RestoreTicketRpc restoreTicketRpc = RestoreTicketRpc(db);
+    final DeleteAllTicketsRpc deleteAllTicketsRpc = DeleteAllTicketsRpc(db);
+    final AddAllTicketsRpc saveAllTicketsRpc = AddAllTicketsRpc(db);
+    final DeleteTicketRpc deleteTicketRpc = DeleteTicketRpc(db);
+
+    return TicketsService(
+      createTicketRpc,
+      disableTicketRpc,
+      getTicketsRpc,
+      restoreTicketRpc,
+      saveAllTicketsRpc,
+      deleteAllTicketsRpc,
+      deleteTicketRpc,
+    );
+  }
+}
+
+class TicketsServiceNoSembast implements TicketsServiceAbstract {
+  const TicketsServiceNoSembast();
+
+  @override
+  get addAllTicketsRpc => AddAllTicketsFakeRpc();
+
+  @override
+  get addTicketRpc => AddTicketFakeRpc();
+
+  @override
+  get deleteAllTicketsRpc => DeleteAllTicketsFakeRpc();
+
+  @override
+  get deleteTicketRpc => DeleteTicketFakeRpc();
+
+  @override
+  get disableTicketRpc => DisableTicketFakeRpc();
+
+  @override
+  get getAllTicketsRpc => GetAllTicketsFakeRpc();
+
+  @override
+  get restoreTicketRpc => RestoreTicketFakeRpc();
 }